+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/WebAppQueryId.kt
@@ -1,0 +1,10 @@
+package dev.inmo.tgbotapi.types
+
+import kotlinx.serialization.Serializable
+import kotlin.jvm.JvmInline
+
+@Serializable
+@JvmInline
+value class WebAppQueryId(
+    val string: String
+)