+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * `InlineQueryIdentifier` now is `value class`
   * `MediaGroupIdentifier` now is `value class`
   * `CallbackQueryIdentifier` now is `value class`
+  * `WebAppQueryId` now is `value class`
 
 ## 10.1.2